requirements.txt
streamlit
openai-agents
python-dotenv
pydantic
tailwind-styled-components
